requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
h5py==2.9.0
html5lib==1.0.1
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==4.10.1
ipython==5.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.0
numpy==1.16.5
pandas==0.23.4
pandocfilters==1.4.2
path.py==11.5.2
pathlib2==2.3.5
python-dateutil==2.7.5
pytz==2018.9
redis==3.0.1
requests==2.21.0
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.3
six==1.12.0
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-hub==0.7.0
urllib3==1.24.1
Werkzeug==0.16.0
wincertstore==0.2